Migrate raids module to new structure
